requirements-macos-pinned.txt
# macOS Apple Silicon (M1/M2/M3) pinned requirements
# For Python 3.11 - based on working environment

# 1. NumPy (compatible version for macOS and whisperx 3.4.3)
numpy>=2.0.2,<2.1

# 2. PyTorch ecosystem (with MPS support, works with numpy 2.x at runtime)
torch==2.8.0
torchaudio==2.8.0

# 3. Core ML/Audio dependencies
onnxruntime==1.19.2
av==15.1.0
numba==0.60.0

# 4. PyAnnote ecosystem (install BEFORE whisperx)
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1

# 5. Whisper ecosystem
openai-whisper==20231117
faster-whisper==1.2.0
ctranslate2==4.4.0

# 6. WhisperX
whisperx==3.4.3

# 7. HuggingFace ecosystem
huggingface_hub==0.35.3
transformers==4.57.0
tokenizers==0.22.1

# 8. Audio processing
librosa==0.10.2.post1
soundfile==0.12.1
audioread==3.0.1

# 9. Metadata and subtitles
tmdbsimple==2.9.1
tmdbv3api==1.9.0
pysubs2==1.7.3

# 10. NER and NLP
spacy==3.8.2
rapidfuzz==3.10.1

# 11. UI and utilities
rich==13.9.4
python-dotenv==1.0.1

# 12. System utilities
psutil==6.1.0

# 13. Additional utilities
pyyaml==6.0.2
requests==2.32.3
python-json-logger==3.2.1
pydantic==2.10.3
pydantic-settings==2.6.1

# 14. Supporting dependencies
certifi>=2024.0.0
charset-normalizer>=3.0.0
idna>=3.0
urllib3>=2.0.0
